refactor: rename incorrect namespaces in tests
