+++ tests/NetDaemon.Daemon.Tests/Fluent/FluentTests.cs
@@ -1,13 +1,13 @@
-﻿using Microsoft.Extensions.Logging;
-using Moq;
-using System;
+﻿using System;
 using System.Dynamic;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
+using Moq;
 using NetDaemon.Common;
 using Xunit;
 
-namespace NetDaemon.Daemon.Tests
+namespace NetDaemon.Daemon.Tests.Fluent
 {
     /// <summary>
     ///     Tests the fluent API parts of the daemon
