+++ tests/NetDaemon.Daemon.Tests/Reactive/RxAppTest.cs
@@ -1,13 +1,13 @@
-using Moq;
 using System;
 using System.Dynamic;
 using System.Linq;
 using System.Reactive.Linq;
 using System.Threading.Tasks;
+using Moq;
 using NetDaemon.Common.Reactive;
 using Xunit;
 
-namespace NetDaemon.Daemon.Tests
+namespace NetDaemon.Daemon.Tests.Reactive
 {
     /// <summary>
     ///     Tests the fluent API parts of the daemon
